refactor(layout.tsx): remove unused imports and code feat(layout.tsx): add AuthProvider and Nav components refactor(page.tsx): remove unused code feat(AuthProvider.tsx): create AuthProvider component feat(Nav.tsx): create Nav component refactor([...nextauth].ts): remove console.log statement
