+++ components/index.ts
@@ -1,3 +1,4 @@
 export { default as Ping } from './Ping';
+export { default as Nav } from './Nav/Nav';
 export * from './Schedule';
 export * from './FitnessClasses';
